Use Elasticsearch cluster health for check instead of ping
Ping will only allow to return healthy or unhealthy statuses. By using
Elasticsearch Cluster Health API [1] we can also return degraded if
the cluster is in yellow state.

Note that I needed to convert the cluster status to a string instead
of using Elasticsearch.Net.Health enum, as otherwise we would need to
add a reference to the Elasticsearch.Net NuGet package.

1: https://www.elastic.co/guide/en/elasticsearch/reference/current/cluster-health.html