Making "safe" defaults for handling connections
Adding an InternalRabbitMQHealthCheck that is only used in cases where we know that the health check instance will be reused instead of creating a new RabbitMQHealthCheck for each health evaluation occurance.

Updating all extension methods to ensure that a single InternalRabbitMQHealthCheck instance is used for each registration instead of having a new instance created for each health evaluation.