Bumping apiVersion for healtchecks crd definition
When deploying the healthchecks operator on the latest version of `kind` it will fail, because `apiextensions.k8s.io/v1beta1` is not supported any longer.
This bumps the `apiVersion` to the stable version and adds minimal changes to adhere to the specification.

Also adding a copy of the pre v1.22 CRD.