Fixes ServiceBus PeekMessageAsync Cancellation Token bug
ServiceBusReceiver.PeekMessageAsync does not necessarily respect the
cancellation token.
See https://github.com/Azure/azure-sdk-for-net/issues/52620.
To get around this limitation, we create a task with the cancellation
token and use WhenAny to return on cancel even if the receiver hasn't
returned.
We also add tests to validate the correct behavior.